Split install into macos specific install and generic install. Fix create_file_link

+++ install.sh
@@ -19,7 +19,7 @@ create_file_link() {
   local SOURCE=$1
   local DEST=$2
 
-  if [ -f "${DEST}" ]; then
+  if [ -f "${DEST}" ] && [ ! -L "${DEST}" ]; then
     backup_file $DEST
     rm $DEST
   fi
@@ -49,25 +49,33 @@ get_os() {
   uname -s
 }
 
-# needed for having the app switcher show on the monitor that the
-# dock was shown on lasa.
-if [ "$(get_os)" == "Darwin" ]; then
+general_install() {
+  if app_exists nvim; then
+    create_dir_link $PWD/nvim $HOME/.config/nvim
+    create_dir_link $PWD/vimwiki $HOME/vimwiki
+  fi
+
+  if app_exists kitty; then
+    create_dir_link $PWD/kitty $HOME/.config/kitty
+  fi
+
+  if app_exists tmux; then
+    create_file_link $PWD/tmux/.tmux.conf $HOME/.tmux.conf
+  fi
+}
+
+macos_specific_install() {
+  # needed for having the app switcher show on the monitor that the
+  # dock was shown on lasa.
   defaults write com.apple.Dock appswitcher-all-displays -bool true
+
   create_link $PWD/bin $HOME/bin
+  create_file_link $PWD/git/.gitconfig $HOME/.gitconfig
+  create_dir_link $PWD/git/.git-template $HOME/.git-template
+}
+
+if [ "$(get_os)" == "Darwin" ]; then
+  macos_specific_install
 fi
 
-if app_exists nvim; then
-  create_dir_link $PWD/nvim $HOME/.config/nvim
-  create_dir_link $PWD/vimwiki $HOME/vimwiki
-fi
-
-if app_exists kitty; then
-  create_dir_link $PWD/kitty $HOME/.config/kitty
-fi
-
-if app_exists tmux; then
-  create_file_link $PWD/tmux/.tmux.conf $HOME/.tmux.conf
-fi
-
-create_file_link $PWD/git/.gitconfig $HOME/.gitconfig
-create_dir_link $PWD/git/.git-template $HOME/.git-template
+general_install
